Package rename and moved methods to service
