+++ src/main/java/cz/foot/SpringBootFootballApp.java
@@ -1,0 +1,29 @@
+package cz.foot;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@SpringBootApplication
+public class SpringBootFootballApp {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SpringBootFootballApp.class, args);
+	}
+
+
+	@Bean
+	public WebMvcConfigurer corsConfigurer() {
+		return new WebMvcConfigurer() {
+			@Override
+			public void addCorsMappings(CorsRegistry registry) {
+				registry.addMapping("/**").allowedOrigins("http://localhost:3000").allowedMethods("*");
+
+			}
+		};
+	}
+
+
+}
