+++ src/main/java/cz/foot/model/League.java
@@ -1,0 +1,57 @@
+package cz.foot.model;
+
+import javax.persistence.*;
+import java.util.List;
+
+@Entity
+
+public class League {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String name;
+
+    @ManyToMany
+    @JoinTable(
+            name = "teams_in_league",
+            joinColumns = @JoinColumn(name = "league_id"),
+            inverseJoinColumns = @JoinColumn(name="team_id"))
+
+    private List<Team> teams;
+
+    public League() {
+    }
+
+    public League(String name, List<Team> teams) {
+        this.name = name;
+        this.teams = teams;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public List<Team> getTeams() {
+        return teams;
+    }
+
+    public void setTeams(List<Team> teams) {
+        this.teams = teams;
+    }
+
+    public void addTeam(Team team) {
+        teams.add(team);
+    }
+}
