+++ src/main/java/cz/foot/model/Position.java
@@ -1,0 +1,6 @@
+package cz.foot.model;
+
+public enum Position {
+    STRIKER, MIDFIELD, DEFENDER, GOALKEEPER
+
+}
