+++ src/main/java/cz/foot/model/StaffMember.java
@@ -1,0 +1,54 @@
+package cz.foot.model;
+
+import javax.persistence.*;
+
+@Entity
+public class StaffMember {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+    private String name;
+    private String surname;
+    private String role;
+    @ManyToOne
+    @JoinColumn(name = "ts_fk", insertable = false, updatable = false)
+    private Team team;
+
+
+    public StaffMember() {
+    }
+
+
+
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getSurname() {
+        return surname;
+    }
+
+    public void setSurname(String surname) {
+        this.surname = surname;
+    }
+
+    public String getRole() {
+        return role;
+    }
+
+    public void setRole(String role) {
+        this.role = role;
+    }
+}
