+++ src/main/java/cz/foot/repository/LeagueRepository.java
@@ -1,0 +1,9 @@
+package cz.foot.repository;
+
+import cz.foot.model.League;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface LeagueRepository extends CrudRepository<League,Long>{
+}
