+++ src/main/java/cz/foot/repository/PlayerRepository.java
@@ -1,0 +1,9 @@
+package cz.foot.repository;
+
+import cz.foot.model.Player;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PlayerRepository extends CrudRepository<Player,Long>{
+}
