+++ src/main/java/cz/foot/repository/TeamRepository.java
@@ -1,0 +1,11 @@
+package cz.foot.repository;
+
+
+
+import cz.foot.model.Team;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TeamRepository extends CrudRepository<Team,Long> {
+}
