+++ src/main/java/cz/foot/service/JwtUserDetailsService.java
@@ -1,0 +1,25 @@
+
+package cz.foot.service;
+
+import java.util.ArrayList;
+
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+public class JwtUserDetailsService implements UserDetailsService {
+
+	@Override
+	public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+		if ("admin".equals(username)) {
+			return new User("admin", "$2a$10$bcc2hqTLaedpm.o02KsowOwVng9rhxQI2gQHBGKOb5uu/L5RnNUBa",
+					new ArrayList<>());
+		} else {
+			throw new UsernameNotFoundException("User not found with username: " + username);
+		}
+	}
+
+}
