--- src/main/java/cz/osu/prf/kip/config/JwtRequestFilter.java
@@ -1,76 +1,0 @@
-package cz.osu.prf.kip.config;
-
-
-import java.io.IOException;
-
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import cz.osu.prf.kip.service.JwtUserDetailsService;
-
-import io.jsonwebtoken.ExpiredJwtException;
-
-@Component
-public class JwtRequestFilter extends OncePerRequestFilter {
-
-	@Autowired
-	private JwtUserDetailsService jwtUserDetailsService;
-
-	@Autowired
-	private JwtTokenUtil jwtTokenUtil;
-
-
-
-
-	@Override
-	protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain chain)
-			throws ServletException, IOException {
-
-		final String requestTokenHeader = request.getHeader("Authorization");
-
-		String username = null;
-		String jwtToken = null;
-		// JWT Token poslán ve formátu "Bearer token". Odstraní slovo Bearer.
-		if (requestTokenHeader != null && requestTokenHeader.startsWith("Bearer ")) {
-			jwtToken = requestTokenHeader.substring(7);
-			try {
-				username = jwtTokenUtil.getUsernameFromToken(jwtToken);
-				System.out.println("Jméno uživatele u requestu je " + username);
-			} catch (IllegalArgumentException e) {
-				System.out.println("Request neobsahuje JWT token");
-			} catch (ExpiredJwtException e) {
-				System.out.println("JWT Tokenu skončila platnost");
-			}
-		} else {
-			logger.warn("JWT Token nezacina slovem Bearer, nebo request není autorizovaný");
-		}
-
-		//Validuj token
-		if (username != null && SecurityContextHolder.getContext().getAuthentication() == null) {
-
-			UserDetails userDetails = this.jwtUserDetailsService.loadUserByUsername(username);
-
-			// Token je validní, vytvoře security realm pro Spring Security
-			if (jwtTokenUtil.validateToken(jwtToken, userDetails)) {
-
-				UsernamePasswordAuthenticationToken usernamePasswordAuthenticationToken = new UsernamePasswordAuthenticationToken(
-						userDetails, null, userDetails.getAuthorities());
-				usernamePasswordAuthenticationToken
-						.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
-				SecurityContextHolder.getContext().setAuthentication(usernamePasswordAuthenticationToken);
-			}
-		}
-		chain.doFilter(request, response);
-	}
-
-}
