--- src/main/java/cz/osu/prf/kip/model/Position.java
@@ -1,6 +1,0 @@
-package cz.osu.prf.kip.model;
-
-public enum Position {
-    STRIKER, MIDFIELD, DEFENDER, GOALKEEPER
-
-}
