--- src/main/java/cz/osu/prf/kip/repository/LeagueRepository.java
@@ -1,9 +1,0 @@
-package cz.osu.prf.kip.repository;
-
-import cz.osu.prf.kip.model.League;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface LeagueRepository extends CrudRepository<League,Long>{
-}
