--- src/main/java/cz/osu/prf/kip/repository/PlayerRepository.java
@@ -1,9 +1,0 @@
-package cz.osu.prf.kip.repository;
-
-import cz.osu.prf.kip.model.Player;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PlayerRepository extends CrudRepository<Player,Long>{
-}
