+++ src/main/resources/application.properties
@@ -1,3 +1,6 @@
+
+
+
 jwt.secret=7VBAP
 jwt.get.token.uri=/authenticate
